ci: added test all and conflict check

+++ contests/Weekly Contest/515/Q1.cpp
@@ -9,7 +9,7 @@ class Solution {
   int nearestDrone(vector<vector<int>>& drones, vector<int>& target) {
     int min = INT_MAX, tmp, idx = -1;
 
-    for (int i = 0; i < drones.size(); ++i) {
+    for (int i = 0; i < (int)drones.size(); ++i) {
       tmp = abs(target[0] - drones[i][0]) + abs(target[1] - drones[i][1]);
       if (tmp > drones[i][2]) continue;
