+++ problems/0-999/344.cpp
@@ -6,7 +6,7 @@ class Solution {
  public:
   void reverseString(vector<char>& s) {
     vector<char> s1(s.size(), 0);
-    for (int i = 0; i < s.size(); ++i) {
+    for (int i = 0; i < (int)s.size(); ++i) {
       s1[i] = s[s.size() - i - 1];
     }
