+++ problems/1000-1999/1365.cpp
@@ -7,7 +7,7 @@ class Solution {
  public:
   vector<int> smallerNumbersThanCurrent(vector<int>& nums) {
     vector<int> ans(nums.size(), 0);
-    for (int i = 0; i < nums.size(); ++i) {
+    for (int i = 0; i < (int)nums.size(); ++i) {
       for (int j = 0; j < i; ++j) {
         if (nums[i] < nums[j]) ans[j]++;
         if (nums[i] > nums[j]) ans[i]++;
