+++ problems/1000-1999/1512.cpp
@@ -7,8 +7,8 @@ class Solution {
  public:
   int numIdenticalPairs(vector<int>& nums) {
     int count = 0;
-    for (int i = 0; i < nums.size(); i++) {
-      for (int j = i + 1; j < nums.size(); j++) {
+    for (int i = 0; i < (int)nums.size(); i++) {
+      for (int j = i + 1; j < (int)nums.size(); j++) {
         if (nums[i] == nums[j]) {
           count++;
         }
