+++ problems/1000-1999/1672.cpp
@@ -6,9 +6,9 @@ class Solution {
  public:
   int maximumWealth(vector<vector<int>>& accounts) {
     int ans = 0;
-    for (int i = 0; i < accounts.size(); ++i) {
+    for (int i = 0; i < (int)accounts.size(); ++i) {
       int sum = 0;
-      for (int j = 0; j < accounts[i].size(); ++j) {
+      for (int j = 0; j < (int)accounts[i].size(); ++j) {
         sum += accounts[i][j];
       }
