+++ problems/1000-1999/1720.cpp
@@ -7,7 +7,7 @@ class Solution {
  public:
   vector<int> decode(vector<int>& encoded, int first) {
     vector<int> result = {first};
-    for (int i = 0; i < encoded.size(); ++i) {
+    for (int i = 0; i < (int)encoded.size(); ++i) {
       result.push_back(result.back() ^ encoded[i]);
     }
     return result;
