+++ problems/1000-1999/1920.cpp
@@ -7,7 +7,7 @@ class Solution {
  public:
   vector<int> buildArray(vector<int>& nums) {
     vector<int> ans(nums.size());
-    for (int i = 0; i < nums.size(); ++i) {
+    for (int i = 0; i < (int)nums.size(); ++i) {
       ans[i] = nums[nums[i]];
     }
