+++ problems/2000-2999/2486.cpp
@@ -6,12 +6,12 @@ class Solution {
  public:
   int appendCharacters(string s, string t) {
     int j = 0;
-    for (int i = 0; i < s.size(); ++i) {
+    for (int i = 0; i < (int)s.size(); ++i) {
       if (s[i] == t[j]) {
         j++;
       }
 
-      if (j == t.size()) break;
+      if (j == (int)t.size()) break;
     }
 
     return t.size() - j;
