+++ problems/2000-2999/2904.cpp
@@ -7,11 +7,11 @@ class Solution {
  public:
   string shortestBeautifulSubstring(string s, int k) {
     string smallest = "";
-    for (int i = 0; i < s.length(); ++i) {
+    for (int i = 0; i < (int)s.length(); ++i) {
       if (s[i] == '1') {
         int cnt = 0;
         string beauty = "";
-        for (int j = i; j < s.length(); ++j) {
+        for (int j = i; j < (int)s.length(); ++j) {
           beauty.push_back(s[j]);
           if (s[j] == '1') cnt++;
           if (cnt == k) break;
