+++ problems/3000-3999/3146.cpp
@@ -7,8 +7,8 @@ class Solution {
  public:
   int findPermutationDifference(string s, string t) {
     int diff = 0;
-    for (int i = 0; i < s.size(); ++i) {
-      for (int j = 0; j < t.size(); ++j) {
+    for (int i = 0; i < (int)s.size(); ++i) {
+      for (int j = 0; j < (int)t.size(); ++j) {
         if (s[i] == t[j]) {
           diff += abs(i - j);
           break;
