+++ problems/3000-3999/3264.cpp
@@ -8,7 +8,7 @@ class Solution {
   vector<int> getFinalState(vector<int>& nums, int k, int multiplier) {
     for (int i = 0; i < k; ++i) {
       int minIndex = 0;
-      for (int j = 1; j < nums.size(); ++j) {
+      for (int j = 1; j < (int)nums.size(); ++j) {
         if (nums[j] < nums[minIndex]) {
           minIndex = j;
         }
