+++ problems/3000-3999/3498.cpp
@@ -7,7 +7,7 @@ class Solution {
   int reverseDegree(string s) {
     int ans = 0;
 
-    for (int i = 1; i <= s.length(); ++i) {
+    for (int i = 1; i <= (int)s.length(); ++i) {
       ans += i * (26 - s[i - 1] + 'a');
     }
