+++ problems/3000-3999/3668.cpp
@@ -8,7 +8,7 @@ class Solution {
  public:
   vector<int> recoverOrder(vector<int>& order, vector<int>& friends) {
     vector<int> result;
-    for (int i = 0; i < order.size(); ++i) {
+    for (int i = 0; i < (int)order.size(); ++i) {
       if (find(friends.begin(), friends.end(), order[i]) != friends.end()) {
         result.push_back(order[i]);
       }
