+++ problems/3000-3999/3931.cpp
@@ -7,7 +7,7 @@ using namespace std;
 class Solution {
  public:
   bool isAdjacentDiffAtMostTwo(string s) {
-    for (int i = 1; i < s.size(); ++i) {
+    for (int i = 1; i < (int)s.size(); ++i) {
       if (abs(s[i] - s[i - 1]) > 2) return false;
     }
     return true;
