+++ problems/4000-4999/4025.cpp
@@ -6,7 +6,7 @@ class Solution {
  public:
   int minPenalty(int period, vector<int>& lights, vector<int>& arrivalTime) {
     int max = 0;
-    for (int i = 0; i < lights.size(); ++i) {
+    for (int i = 0; i < (int)lights.size(); ++i) {
       if (max < lights[i]) max = lights[i];
     }
